requirements.txt
praw==7.7.1
Pillow==10.1.0
requests==2.31.0 